requirements.txt
Django==1.9.5
djangorestframework==3.3.3
Pillow==3.2.0
wheel==0.24.0
